feat: update CommentController to use Lombok for constructor injection, add authorization checks for comment updates and deletions, and improve error handling in deleteById method; modify CommentDto and CreateCommentRequest to use Long for postId

+++ src/main/java/org/amalitech/dtos/CreateCommentRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CreateCommentRequest {
-    private int postId;
+    private Long postId;
     private String username;
     private String body;
     private String parentCommentId;
